refactor project structure for recursive implementation of trees
